fix: Avoiding exceptions loading files on android

+++ src/Uno.Extensions.Storage.UI/FileStorage.cs
@@ -3,17 +3,34 @@ namespace Uno.Extensions.Storage;
 
 public class FileStorage : IStorage
 {
-	public async Task<string> CreateLocalFolderAsync(string foldername)
+	private async Task<bool> FileExistsInPackage(string filename)
+	{
+#if __ANDROID__
+		var assets = global::Android.App.Application.Context.Assets;
+		var files = assets?.List("");
+		filename = Path.GetFileNameWithoutExtension(filename).Replace('.', '_') + Path.GetExtension(filename);
+		return files?.Contains(filename)??false;
+#else
+		return true;
+#endif
+
+	}
+
+	public async Task<string> CreateFolderAsync(string foldername)
 	{
 		var localFolder = Windows.Storage.ApplicationData.Current.LocalFolder;
 		var folder = await localFolder.CreateFolderAsync(foldername, CreationCollisionOption.OpenIfExists);
 		return folder.Path;
 	}
 
-	public async Task<string?> ReadFileAsync(string filename)
+	public async Task<string?> ReadPackageFileAsync(string filename)
 	{
 		try
 		{
+			if(!await FileExistsInPackage(filename))
+			{
+				return default;
+			}
 			var storageFile = await StorageFile.GetFileFromApplicationUriAsync(new Uri($"ms-appx:///{filename}"));
 			if (File.Exists(storageFile.Path))
 			{
@@ -21,27 +38,33 @@ public class FileStorage : IStorage
 				return settings;
 			}
 
-			return null;
+			return default;
 		}
 		catch
 		{
-			return null;
+			return default;
 		}
 
 	}
 
-	public async Task<Stream> OpenFileAsync(string filename)
+	public async Task<Stream?> OpenPackageFileAsync(string filename)
 	{
+		if (!await FileExistsInPackage(filename))
+		{
+			return default;
+		}
+
 		var storageFile = await StorageFile.GetFileFromApplicationUriAsync(new Uri($"ms-appx:///{filename}"));
 		var stream = await storageFile.OpenStreamForReadAsync();
 		return stream;
 	}
 
-	public async Task WriteFileAsync(string filename, string text)
+	public async Task WriteFileAsync(string filename, string text, bool overwrite)
 	{
-		//var localFolder = Windows.Storage.ApplicationData.Current.LocalFolder;
-		//var settingsFile = await localFolder.CreateFileAsync($"{filename}", CreationCollisionOption.OpenIfExists);
-		File.WriteAllText(filename, text);
+		if (!File.Exists(filename) || overwrite)
+		{
+			File.WriteAllText(filename, text);
+		}
 	}
 
 }
